Moved ControllerSocket and DataStream into their own files
